Segmentize outgoing packets.

+++ utcp.h
@@ -48,7 +48,7 @@ extern struct utcp *utcp_init(utcp_accept_t accept, utcp_pre_accept_t pre_accept
 extern void utcp_exit(struct utcp *utcp);
 extern struct utcp_connection *utcp_connect(struct utcp *utcp, uint16_t port, utcp_recv_t recv, void *priv);
 extern void utcp_accept(struct utcp_connection *utcp, utcp_recv_t recv, void *priv);
-extern int utcp_send(struct utcp_connection *connection, void *data, size_t len);
+extern ssize_t utcp_send(struct utcp_connection *connection, void *data, size_t len);
 extern int utcp_recv(struct utcp *utcp, void *data, size_t len);
 extern int utcp_close(struct utcp_connection *connection);
 extern int utcp_abort(struct utcp_connection *connection);
